Добавляет фото на сайт
