+++ index.html
@@ -6,9 +6,10 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="/js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
   <script type="module" src="/js/generateRandom.js"></script>
   <script type="module" src="/js/comment.js"></script>
+  <script type="module" src="/js/picture.js"></script>
   <title>Кекстаграм</title>
 </head>
